jnipap: Exception classes in individual files
Split up the jnipap exceptions into individual files. This fixes the warning

  warning: auxiliary class ... in ... should not be accessed from outside its own source file

which is emitted when building the test cases.
