+++ jnipap/jnipap/DuplicateException.java
@@ -1,0 +1,21 @@
+package jnipap;
+
+/**
+ * The requested addition/change viloates unique constraints
+ *
+ * For example, when creating a VRF which already exists.
+ */
+class DuplicateException extends JnipapException {
+
+	// version ID for serialization
+	private static final long serialVersionUID = 0L;
+
+	public DuplicateException(String msg) {
+		super(msg);
+	}
+
+	public DuplicateException(Throwable cause) {
+		super(cause);
+	}
+
+}
